+++ src/elements/conditions/sendouts/SendoutScheduleCondition.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\campaign\elements\conditions\sendouts;
+
+use craft\elements\conditions\ElementCondition;
+
+/**
+ * @since 2.0.0
+ */
+class SendoutScheduleCondition extends ElementCondition
+{
+    /**
+     * @inheritdoc
+     */
+    protected function conditionRuleTypes(): array
+    {
+        return [];
+    }
+}